Fix TD packet IDs and the oldest-packet search
TD's packet ID and last sequence/read IDs were unsigned long, 64 bits on
Linux, so the 0xffffffff "nothing received" value plus one never wrapped to
the first packet: acknowledged packets were never delivered. They are
uint32_t again, as in the original and in RA, which restores TD's 8-byte
packet header.

The oldest unacknowledged packet search in both games started from a
0xffffffff time, which every tick exceeds after about 828 days of uptime.
It moves into ConnectionClass::OldestUnackedSend, which tracks whether it
found a packet. New tests cover delivery order, duplicate resends and tick
values past 2^32.

+++ src/ra/combuf.cc
@@ -867,7 +867,7 @@ void CommBufferClass::Mono_Debug_Print(int refresh) {
   struct CommHdr {  // this mirrors the CommHeaderType
     unsigned short MagicNumber;
     unsigned char Code;
-    unsigned long PacketID;
+    uint32_t PacketID;
   }* hdr;
 
   //------------------------------------------------------------------------
@@ -1043,7 +1043,7 @@ void CommBufferClass::Mono_Debug_Print2(int refresh) {
   struct CommHdr {  // this mirrors the CommHeaderType
     unsigned short MagicNumber;
     unsigned char Code;
-    unsigned long PacketID;
+    uint32_t PacketID;
   }* hdr;
 
   //------------------------------------------------------------------------
