+++ src/td/noseqcon.h
@@ -113,23 +113,23 @@ class NonSequencedConnClass : public ConnectionClass {
   Running totals of # of packets we send & receive which require an ACK,
   and those that don't.
   .....................................................................*/
-  unsigned long NumRecNoAck = 0;
-  unsigned long NumRecAck = 0;
-  unsigned long NumSendNoAck = 0;
-  unsigned long NumSendAck = 0;
+  uint32_t NumRecNoAck = 0;
+  uint32_t NumRecAck = 0;
+  uint32_t NumSendNoAck = 0;
+  uint32_t NumSendAck = 0;
 
   /*.....................................................................
   This is the ID of the last consecutively-received packet; anything older
   than this, we know is a resend.  Anything newer than this MUST be lying
   around in the Queue for us to detect it as a resend.
   .....................................................................*/
-  unsigned long LastSeqID = 0xffffffff;
+  uint32_t LastSeqID = 0xffffffff;  // 0xffffffff: none yet; +1 wraps to 0
 
   /*.....................................................................
   This is the ID of the PACKET_DATA_ACK packet we read last; it ensures
   that the application reads that type of packet in order.
   .....................................................................*/
-  unsigned long LastReadID = 0xffffffff;
+  uint32_t LastReadID = 0xffffffff;  // 0xffffffff: none yet; +1 wraps to 0
 };
 
 #endif  // CNC_RED_ALERT_TD_NOSEQCON_H_
